+++ src/components/scenes/SettingsScene.tsx
@@ -418,6 +418,10 @@ export const SettingsScene: React.FC<Props> = props => {
     navigation.navigate('swapSettings')
   })
 
+  const handleOpenDebugSettings = useHandler((): void => {
+    navigation.navigate('debugSettings')
+  })
+
   const handleSpendingLimits = useHandler(async (): Promise<void> => {
     if (await hasLock()) return
     navigation.navigate('spendingLimits')
@@ -745,6 +749,12 @@ export const SettingsScene: React.FC<Props> = props => {
                 onPress={() => {
                   navigation.navigate('giftCardAccountInfo', {})
                 }}
+              />
+            )}
+            {developerModeOn && (
+              <SettingsTappableRow
+                label={lstrings.settings_debug_title}
+                onPress={handleOpenDebugSettings}
               />
             )}
           </EdgeCard>